requirements.txt
PyMuPDF
camelot-py[cv]
pandas
numpy